쪽지보내기, 답장하기 컴포넌트 분리
